This update reflects the learnings that came from setting up a new Raspberry pi node from scratch to act as a digipeater and power-control node.  The pi was setup with Raspbian Jessie Lite, which then mandated that the 'remote-access.yml' playbook also installs python-pip and git.
Also, we added an 'aprx' role that downloads, compiles, and installs
APRX 2.9.0.  Previously, APRX was installed by hand on the digipeater node.

README.md was added and updated to give instructions for setting up a new
node.  There's probably some work yet to do to make it (and the playbook
structure) more comprehensible.